feat: Saga関連のリファクタリングと修正
- `saga_handler.rs`: ページネーションの計算を修正し、型を明示的に指定
- `mod.rs`: 不要なエクスポートを削除
- `main.rs`: ページネーションの計算を改善し、ページとページサイズの最小値を設定
- `execute_saga.rs`: モックのメソッド引数を簡略化
- `start_saga.rs`: モックリポジトリの変数宣言を修正

+++ regions/system/server/rust/saga/src/adapter/handler/saga_handler.rs
@@ -211,7 +211,7 @@ pub async fn list_sagas(
         })
         .collect();
 
-    let has_next = (query.page * query.page_size) < total as i32;
+    let has_next = (query.page as i64 * query.page_size as i64) < total;
 
     Ok(Json(ListSagasResponse {
         sagas: saga_responses,
